% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdGBH.R
\name{tdGBH}
\alias{tdGBH}
\title{Two-Dimensional Group Benjamini-Hochberg (tdGBH) Procedure}
\usage{
tdGBH(
  p.mat,
  pi0.method = c("storey", "lsl", "tst"),
  global.pi0.method = c("storey", "lsl", "tst"),
  shrinkage.method = c("linear", "power"),
  shrinkage.factor = 0.1
)
}
\arguments{
\item{p.mat}{A matrix representing p-values with rows and columns corresponding to the two grouping directions (e.g. genes by cell types).}

\item{pi0.method}{A character string of 'lsl', 'tst', or 'storey', specifying the method for estimating the group-specific null proportions. The default is 'storey'.}

\item{global.pi0.method}{A character string of 'lsl', 'tst', or 'storey', specifying the method for estimating the global null proportion. The default is 'storey'.}

\item{shrinkage.method}{A character string of 'linear' and 'power', indicating whether the shrinkage is on the  linear or power scale.}

\item{shrinkage.factor}{A numeric value between 0 and 1. The factor determines the degree of shrinkage to the global null proportion estimate (0: no shrinkage, 1: maximum shrinkage, i.e. using the global estimate). The default is 0.1.}
}
\value{
Returns the FDR-adjusted p-values.
}
\description{
Two-Dimensional Group Benjamini-Hochberg (tdGBH) Procedure is a multiple testing procedure that offers enhanced power over traditional FDR control methods when the data presents two-way grouping structure. It reweights the p-values based on the informativeness of the respective grouping directions. It requires a matrix of p-values as input with rows and columns corresponding to the two grouping directions (e.g. genes by cell types).
}
\examples{
data(P)
p.adj <- tdGBH(P)
}
\references{
Lu Yang, Pei Wang, Jun Chen. 2dGBH: Two-dimensional Group Benjamini-Hochberg Procedure for False Discovery Rate Control in Two-Way Multiple Testing of Genomic Data
}
\author{
Jun Chen and Lu Yang
}
