% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdGBH.R
\name{tdGBH}
\alias{tdGBH}
\title{Two-Dimensional Group Benjamini-Hochberg (tdGBH) Procedure}
\usage{
tdGBH(
  p.mat,
  pi0.method = "storey",
  global.pi0.method = "storey",
  weight.method = "new",
  shrink = 0.1,
  renorm = F
)
}
\arguments{
\item{p.mat}{A matrix representing p-values where rows correspond to features and columns to outcomes.}

\item{pi0.method}{A character string specifying the method for \link[structSSI]{estimate.pi0}, chosen from 'lsl', 'tst', or 'storey'. The default is 'storey'.}

\item{global.pi0.method}{A character string that indicates the method used for the global \link[structSSI]{estimate.pi0}. Possible values are 'lsl', 'tst', or 'storey'. The default is 'storey'.}

\item{weight.method}{A character string specifying the method for add the weight, chosen from 'geo', 'ari', or 'new', represents geometric mean,  arithmetic mean, and tdgbh method. The default is 'new'.}

\item{shrink}{A numeric value between 0 and 1, serving as a shrinkage factor. It is employed to mitigate the impact of sampling variability. The factor determines the weighted average between global and group-specific proportions of null hypotheses. The default is 0.1.}

\item{renorm}{A logical value indicating whether renormalization should be performed on the weights, which ensures FDR control for non data-driven weights. The default is FALSE.}
}
\value{
Returns the adjusted p-values.
}
\description{
This function applies the two-dimensional group Benjamini-Hochberg (tdGBH) procedure to control the false discovery rate.
}
\examples{
data(P)
p.adj <- tdGBH(P)
}
\references{
Lu Yang, Pei Wang, Jun Chen. 2dGBH: Two-dimensional Group Benjamini-Hochberg Procedure for False Discovery Rate Control in Two-Way Multiple Testing.
}
\author{
Lu Yang, Pei Wang and Jun Chen
}
