% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdGBH.R
\name{tdGBH}
\alias{tdGBH}
\title{tdGBH}
\usage{
tdGBH(p.mat, pi0.method = "storey", global.pi0.method = "storey", shrink = 0.1)
}
\arguments{
\item{p.mat}{a matrix of p values. Rows: features, columns: outcomes.}

\item{pi0.method}{a character string of 'lsl','tst' or 'storey'. The name of any method used in estimate.pi0.}

\item{global.pi0.method}{a character string of 'lsl','tst' or 'storey'. The name of any method used in estimate.pi0.}

\item{shrink}{a number between 0-1. A shrinkage factor to reduce the effects of sampling variation,which is the weighted average of the global and group-specific proportions of null hypotheses.}
}
\value{
The adjusted FDR.
}
\description{
tdGBH
}
